# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/user/signup":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"]}},"/user":{"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["User"],"security":[{"bearer":[]}]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":false,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/User"}}}}},"tags":["User"],"security":[{"bearer":[]}]}},"/auth":{"get":{"operationId":"AuthController_checkUserController","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Auth"],"security":[{"bearer":[]}]}},"/auth/signin":{"post":{"operationId":"AuthController_loginController","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthDto"}}}},"responses":{"201":{"description":""}},"tags":["Auth"]}},"/files":{"post":{"operationId":"FilesController_create","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/CreateFilesDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Files"],"security":[{"bearer":[]}]},"get":{"operationId":"FilesController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"default":1,"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"default":10,"type":"number"}},{"name":"generate_file_name","required":false,"in":"query","schema":{"type":"string"}},{"name":"file_ext","required":false,"in":"query","schema":{"type":"string"}},{"name":"file_storage","required":false,"in":"query","schema":{"type":"number"}},{"name":"file_data","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ResponFilesDto"}}}}},"tags":["Files"],"security":[{"bearer":[]}]}},"/files/{id}":{"get":{"operationId":"FilesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Files"}}}}},"tags":["Files"],"security":[{"bearer":[]}]},"patch":{"operationId":"FilesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/UpdateFilesDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Files"],"security":[{"bearer":[]}]},"delete":{"operationId":"FilesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Files"}}}}},"tags":["Files"],"security":[{"bearer":[]}]}},"/files/files-manual-query":{"post":{"operationId":"FilesController_filesManualQuery","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilesManualQueryDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["Files"],"security":[{"bearer":[]}]}}},"info":{"title":"OPEN-API POS","description":"Documentasi untuk api point of sale","version":"1.3","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"nama_user":{"type":"string","minLength":8,"maxLength":64},"email":{"type":"string","minLength":6,"maxLength":32},"username":{"type":"string","minLength":8,"maxLength":32},"password":{"type":"string","minLength":8,"maxLength":32}},"required":["nama_user","email","username","password"]},"User":{"type":"object","properties":{"id":{"type":"number"},"nama_user":{"type":"string"},"email":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"produk":{"$ref":"#/components/schemas/Files"}},"required":["id","nama_user","email","username","password","create_at","update_at","produk"]},"Files":{"type":"object","properties":{"id":{"type":"number"},"generate_file_name":{"type":"string"},"file_ext":{"type":"string"},"file_storage":{"type":"string"},"file_data":{"type":"string"},"create_at":{"format":"date-time","type":"string"},"update_at":{"format":"date-time","type":"string"},"user":{"$ref":"#/components/schemas/User"}},"required":["id","generate_file_name","file_ext","file_storage","file_data","create_at","update_at","user"]},"UpdateUserDto":{"type":"object","properties":{"id":{"type":"number"},"nama_user":{"type":"string","minLength":8,"maxLength":64},"email":{"type":"string","minLength":6,"maxLength":32},"username":{"type":"string","minLength":8,"maxLength":32},"password":{"type":"string","minLength":8,"maxLength":32}}},"AuthDto":{"type":"object","properties":{"username":{"type":"string","minLength":8,"maxLength":32},"password":{"type":"string","minLength":8,"maxLength":32}},"required":["username","password"]},"CreateFilesDto":{"type":"object","properties":{"file_data":{"type":"string","format":"binary"}},"required":["file_data"]},"FilesDto":{"type":"object","properties":{"id":{"type":"number"},"file_data":{"type":"string","format":"binary"}},"required":["id","file_data"]},"ResponFilesDto":{"type":"object","properties":{"total":{"type":"number"},"pages":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/FilesDto"}}},"required":["total","pages","data"]},"UpdateFilesDto":{"type":"object","properties":{"id":{"type":"number"},"file_data":{"type":"string","format":"binary"}}},"FilesManualQueryDto":{"type":"object","properties":{"condition":{"type":"object","properties":{"generate_file_name":{"required":true,"type":"string"}}},"skip":{"type":"number","default":0},"limit":{"type":"number","default":10},"enable_count":{"type":"number","default":1},"enable_manual_relation_user":{"type":"number","default":1}},"required":["condition"]}}}}